fulfillai/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
httpx>=0.27.0
